% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_rstar_glm.R
\name{create_plots}
\alias{create_plots}
\title{Create Diagnostic Plots for Simulation Results}
\usage{
create_plots(data)
}
\arguments{
\item{data}{Data frame containing the results of the simulation studies.}
}
\value{
A list with ggplot2 objects for the estimate and p-value plots.
}
\description{
This internal function creates diagnostic plots for the results of the simulation studies,
including boxplots for GLM estimates and p-values.
}
\keyword{internal}
