% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LA_rstar.R
\name{LA_rstar}
\alias{LA_rstar}
\title{Launch Shiny App for likelihoodAsy rstar Analysis}
\usage{
LA_rstar()
}
\value{
A Shiny app object that can be run locally.
}
\description{
This function launches a Shiny application that facilitates the setup and execution
of likelihoodAsy rstar analysis. The app allows users to upload a dataset, specify a
model and parameters of interest, and perform the analysis with the option to compute
confidence intervals for r* statistics.
}
\examples{
if (interactive()) {
  LA_rstar()
}

}
\references{
Pierce, D. A., & Bellio, R. (2017). Modern Likelihood-Frequentist Inference.
International Statistical Review / Revue Internationale de Statistique, 85(3),
519–541. \url{doi:10.1111/insr.12232}

Bellio R, Pierce D (2020). likelihoodAsy: Functions for Likelihood Asymptotics.
R package version 0.51, \url{https://CRAN.R-project.org/package=likelihoodAsy}.
}
